+++ src/components/b_navbar.rs
@@ -1,4 +1,3 @@
-use leptos::ev;
 use leptos::html::{Div, A};
 use leptos::*;